Broke up renderer tests into individual files
