+++ apps/site/api/tests/renderer_tests/__init__.py
@@ -1,0 +1,4 @@
+from localground.apps.site.api.tests.renderer_tests.csv_tests import *
+from localground.apps.site.api.tests.renderer_tests.zip_tests import *
+from localground.apps.site.api.tests.renderer_tests.kml_tests import *
+from localground.apps.site.api.tests.renderer_tests.geojson_tests import *
